splitting marginal screening in two
